requirements.txt
boto3==1.24.56
celery==5.2.2
daphne==3.0.2
dj_database_url
Django==3.2.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-environ==0.8.1
django-extensions==3.1.3
django-filter==21.1
django-notifications-hq==1.7.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
factory_boy
Faker==13.7.0
freezegun==1.2.2
mock==4.0.3
mypy==0.931
mypy-extensions==0.4.3
Pillow
pre-commit
psycopg==3.0.3
psycopg2
psycopg2-binary
PyJWT==2.3.0
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
redis==3.5.3
requests==2.26.0
whitenoise==5.3.0
